pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-google-form-complete"
version = "0.1.0"
description = "Psikolojik yaklaşımlara göre Google Form otomatik doldurma botu (9 yaklaşım, 73 soru)"
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "anthropic>=0.40.0",
    "groq>=0.11.0",
    "google-genai>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "responses>=0.25.0",
    "ruff>=0.6.0",
]

[tool.setuptools]
py-modules = ["config", "perspectives", "generator", "submitter", "main"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
